Refactor Controllers plugin
* Use `ArtefactTypes.CONTROLLER` instead of `ControllerArtefactHandler.TYPE`
* Move `Action` `CamelCaseUrlConverter` `HyphenatedUrlConverter` `UrlConverter` from grace-core to grace-web-common
* Relocate `ControllerArtefactHandler` `ControllerArtefactHandlerSpec` to grace-plugin-controllers
* Update `LoggingTransformerSpec` in grace-logging
* Add `ControllerLoggingTransformerSpec` to grace-plugin-controllers
