+++ grace-core/src/main/resources/META-INF/grails.factories
@@ -1,6 +1,5 @@
 grails.core.ArtefactHandler=\
 org.grails.core.artefact.ApplicationArtefactHandler,\
-org.grails.core.artefact.ControllerArtefactHandler,\
 org.grails.core.artefact.ServiceArtefactHandler,\
 org.grails.core.artefact.UrlMappingsArtefactHandler
