+++ grace-plugin-controllers/src/main/resources/META-INF/grails.factories
@@ -1,3 +1,6 @@
+grails.core.ArtefactHandler=\
+org.grails.core.artefact.ControllerArtefactHandler
+
 grails.compiler.traits.TraitInjector=\
 grails.compiler.traits.ControllerTraitInjector
